### 2025-06-05 #8352 @editorialbot
[PRE REVIEW]: OPTIMEO: Optimization Platform for Tuning, Inference, Modeling, Exploration, and Orchestration
**Submitting author:** @colinbousige (<a href="http://orcid.org/0000-0002-0490-2277">Colin Bousige</a>)
**Repository:** https://github.com/colinbousige/OPTIMEO
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
