### 2025-06-05 #8352 @crvernon
To address the paper compile error:

I think the issue is that the `paper.md` needs to specify the date as `30 April 2025` instead of `2025-04-30`
