### 2025-06-12 #8352 @crvernon
:wave: @jbytecode are you able to take this one on since we are wrapping up your other submission?
