### 2025-06-16 #8352 @sgbaird
Happy to review - very relevant for me. I got started with some initial thoughts/impressions of the platform: https://www.youtube.com/playlist?list=PLKFxDV1Aoxg5n1ZyG1H6oYHDkMXPovma6. Will formalize that later
