### 2025-06-29 #8352 @jbytecode
@sgbaird, @VasanthRajendran, @colinbousige - I'm starting the review.

Please follow the instructions in the given issue.
