### 2025-08-18 colinbousige/OPTIMEO#4 @sgbaird
General feedback from sgbaird for JOSS review
based on https://github.com/openjournals/joss-reviews/issues/8510#issuecomment-3055197385

Could you watch https://www.youtube.com/playlist?list=PLKFxDV1Aoxg5n1ZyG1H6oYHDkMXPovma6 and let me know what your reactions/impressions are from that? This is really the "meat" of my feedback.

---

Some other minor notes:

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Could use a bit more fleshing out. https://github.com/colinbousige/OPTIMEO?tab=readme-ov-file#contributing

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I'm not seeing automated tests.

Add a `.gitignore` file. See https://github.com/github/gitignore/blob/main/Python.gitignore for a template that you can copy-paste directly. Otherwise venv will give you 1k+ changes

Consider using `pyproject.toml` rather than `setup.py`.

Related to this, requirements.txt and pyproject.toml contain duplicate info, and web app dependencies aren't well separated from the package dependencies. As a side note, I'm not sure having (or rather, advertising) and additional high-level wrapper is justified. It's fine if they're convenience functions for writing the web app, but it feels like the web app is the main contribution. I have a hard time seeing myself using the wrapper.

---

I'll revisit after hearing back from you.
