### 2025-09-03 colinbousige/OPTIMEO#4 @colinbousige
Hi Sterling,

Many thanks for your feedback!
Following what I got from it, here is a list all the changes I made:

- fixed issue with `ValueError(f"Model was not fit on metric {metric}")` with two objectives and one outcome
- removed the option for Sobol generator in BO: it was an "historic" option I had left before adding the DOE part, it makes indeed more sense the create a DOE, acquire some data, then just run the BO
- made a clearer message if `Nexp>1` and UCB is chosen: it switches back to the default acquisition function (loqEI).
- added a message if the number of outcomes is larger than 2: the web app only handles 2 objectives. In fact, the package version can handle more than 2, but it was a bit of a hassle to make the app handle more than 2 whereas I think the use case is very limited
- also added a message when chosing UCB that "If you don't really know what you are doing, just stick with the default acquisition function and switch this off"
- made the choice of beta in UCB free and not constrained on a slider
- used a clearer terminology: parameters instead of features, objectives instead of metrics, etc.
- un-ignore the `.gitignore` file, and add more stuff to it to ignore the virtual envs
- fixed the web app for optimal designs to compute the minimum number of experiments for it, and thus not return an error since `Nexp` cannot be below this number
- made a change so that loaded data in either BO or ML page stays loaded when switching page
- changed the BO page organization, to show the plot panel below instead of on the right side
- added possibility to remove error bars in the Pareto plot
- added links to display the example notebooks in Google colab
- added automated tests
- switched from `setup.py` to `pyproject.toml`
- updated the Contributing section
- amended the paper to say that the package is just a "bonus", the normal way to use OPTIMEO is with the web app. The BO part of the package is indeed a wrapper for ax, as it was much easier to write the app that way

Concerning your remark on PyPI, I didn't put it there because, as you said, I don't think the main use will be through the wrapper package but rather with the web app. So I didn't see the need to put it there, but I can always do it if you think it's needed.

I hope that I got everything... let me know if you have something else!

Colin
