### 2025-11-19 colinbousige/OPTIMEO#4 @sgbaird
@colinbousige sorry it's taken so long. I think I'm not seeing your latest changes mentioned above, both in the PDF https://github.com/openjournals/joss-reviews/issues/8510#issuecomment-3550594803 and the docs page. Could you double check this?

> * amended the paper to say that the package is just a "bonus", the normal way to use OPTIMEO is with the web app. The BO part of the package is indeed a wrapper for ax, as it was much easier to write the app that way

Minor concern: the following from the manuscript seems somewhat at odds with the above, in terms of differentiating it relative to other packages.

> Finally, these programs are not available as
> 78 standalone Python packages for more advanced use. Such advanced usage might be required
> 79 if the user wants to run the optimization loop in a robotic high-throughput lab, for example –
> 80 which is why this package was developed in the first place.

> * updated the Contributing section

For more than just this one, I'm wondering if I'm seeing the latest changes that you mentioned.

https://colinbousige.github.io/OPTIMEO/optimeo.html?search=contributing

> No search results for 'contributing'.

EDIT: looking through [the commits](https://github.com/colinbousige/OPTIMEO/commits/main/), I'm seeing the changes.
