### 2025-11-19 colinbousige/OPTIMEO#4 @sgbaird
nitpick: I think with most of these methods, the way categorical variables are treated is more or less via random search, not with something intelligent from a categorical space-filling perspective. This can affect the space filling properties. It's a reasonable compromise IMO. Figured I'd at least point out.

I think in the paper you mentioned something about auto-populating the bounds. That only applies for BO if you have existing data, which makes sense.

At first glance, I wasn't really sure what random order does. Thinking a bit, I realized it probably means shuffling the suggested designs.

Not sure what "other (comma separated)" refers to or what I'd put in.

Passing thought: Maybe default number of experiments could be 10 or so? Fine if you decide to leave at one.

Random glitch, I pushed the number of parameters up to 8 or so, and then I rapidly moved back to 1.

```python
AssertionError: This app has encountered an error. The original error message is redacted to prevent data leaks. Full error details have been recorded in the logs (if you're on Streamlit Cloud, click on 'Manage app' in the lower right of your app).
Traceback:
File "/mount/src/doe-doa/pages/1_1._Design_Of_Experiments.py", line 331, in <module>
    doe = DesignOfExperiments(
          ^^^^^^^^^^^^^^^^^^^^
File "/mount/src/doe-doa/optimeo/doe.py", line 155, in __init__
    self.create_design()
File "/mount/src/doe-doa/optimeo/doe.py", line 392, in create_design
    ax_client.create_experiment(
File "/home/adminuser/venv/lib/python3.12/site-packages/ax/service/ax_client.py", line 347, in create_experiment
    experiment = self.make_experiment(
                 ^^^^^^^^^^^^^^^^^^^^^
File "/home/adminuser/venv/lib/python3.12/site-packages/ax/service/utils/instantiation.py", line 945, in make_experiment
    search_space=cls.make_search_space(parameters, parameter_constraints),
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
File "/home/adminuser/venv/lib/python3.12/site-packages/ax/service/utils/instantiation.py", line 741, in make_search_space
    cls.constraint_from_str(c, parameter_map) for c in parameter_constraints
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
File "/home/adminuser/venv/lib/python3.12/site-packages/ax/service/utils/instantiation.py", line 475, in constraint_from_str
    parameter in parameter_names
```

EDIT: just realized this is because I had left in a constraint that depended on two variables, and the error message was actually about the second parameter in the constraint no longer being defined

Good to see measures taken to hide the raw logs. I guess it obscured things a bit in this case.

I guess when you get to a high enough number of parameters, we get a frowny face for some (not all) of the 3D graphs. Not sure what best option is here. Toggle for hiding the plots? Using a plotly dropdown that allows you to select which axes? (I actually somewhat like/prefer having the various combinations displayed all at once).

Passing comment that I'm noticing the run_order is a column, rather than the rows themselves being shuffled. I'm guessing that was intentional.

Someone might be surprised/disappointed if when they reload the web page (Ctrl+R or Ctrl+Shift+R), they lose everything (especially search space in DoE case, and to a lesser extent the BO case if they made significant modifications). Might be a bit frustrating to enter everything again.

Just wanted to ask for confirmation: requesting a batch of new experiments properly handles joint acquisition? In practical terms, uses `get_next_trials()` or `get_next_trial()` in a loop, etc.

> The default value of β is 1, which provides a good balance between exploration and exploitation.

Maybe instead:

> The default value of β is 1, which simplifies to mu(x) + sigma(x), such that the predictions and the model's uncertainty estimates (standard deviations) are balanced equally.

While I find myself wanting to agree, "good" might be a bit subjective.

Minor comment (paper): ax-platform now has [a recent citation](https://github.com/facebook/Ax/issues/281#issuecomment-3550793046) dedicated to the package, albeit based on the v1 release. Might be worth adding next to the 2018 one.

Would it make sense to give people an option to customize/fix a seed? (i.e., still leave by default to random seed, but in case someone wants to fix the seed). Maybe makes sense, maybe not

Nitpick: I can't screenshot it, but if you hover over the + or - here, you get a red circle

Maybe worth adding a tooltip next to or a short blurb beneath "Plot options" related to the slice plot, which people tend to get confused by and I often find difficult to explain succinctly. Maybe a simple narrated example would help

> For example, if your search space is x1 /elem [0, 1], x2 /elem [0,1], x3 /elem [0,1], x4 /elem [0,1], and x5 /elem [0,1], and you want to plot a 2D contour of your model's predictions comparing x1 vs. x2, you'll need to fix the other parameters to a constant value to get a 2D "slice" through the high-dimensional space.

Something unique you could add would be adding transparency to the measured datapoints based on their projected distance from the slice (i.e., the perpendicular distance / shortest distance to the plane). I haven't seen someone plot it like this before, but I think it could help people recognize that the point is far away from the slice.

By default, there are no values present in the boxes for the slice, though I believe ax uses the midpoint of the search space by default. Perhaps worth exposing that in the box by default rather than leaving blank.

Random observation: I was slightly surprised at how long it took to generate a new prediction (~15 s?)

I also recommend including the uncertainty in addition to the average response:

Not sure if I brought up before, but it would be nice to include input fields for objective thresholds with a blurb about their role. By default can be left as inferred, but this is something I often describe to people as a way for them to incorporate their domain knowledge. "If obj_1 is less than X, it doesn't matter how good obj_2 is".

> * added a message if the number of outcomes is larger than 2: the web app only handles 2 objectives. In fact, the package version can handle more than 2, but it was a bit of a hassle to make the app handle more than 2 whereas I think the use case is very limited

I think I'm OK with this because outcome constraints are also supported.
