### 2025-11-19 colinbousige/OPTIMEO#4 @sgbaird
Couple additional remarks:

a problem statement could be added at the beginning here

with a short blurb about the real-world task in the reference in the docs

EDIT: I notice this is present for the BO notebook, just not for the DoE notebook:

I'm mostly thinking about the JOSS checkbox: `Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).`

I think you already have elements of this around, and it might be pretty self-explanatory for the audience that would be landing on the web app. I'll still mark the checkbox, but figured I'd still bring up.
