### 2025-11-19 #8510 @colinbousige
Alright @sgbaird, I think I have answered all your remarks in https://github.com/colinbousige/OPTIMEO/issues/4#issuecomment-3550825029
Many thanks for your review, this really helped me improve the app... and thanks for your positive remarks too!
