### 2025-11-20 #8510 @jbytecode
@sgbaird, @VasanthRajendran, @colinbousige - Since our reviewers checked out all of the review items in their checklists, I will start my editorial stuff soon.

@VasanthRajendran didn't post a report but regarding to their checklist I assume they're happy with the latest form of the submission, right? Any issues?
