+++ paper/paper.bib
@@ -184,10 +184,11 @@ url={https://openreview.net/forum?id=U1f6wHtG1g}
 
 @misc{_statease_2025,
   title = {Statease/Dexpy},
+  author = {{Statease}},
   year = {2025},
   month = mar,
-  urldate = {2025-05-05},
-  howpublished = {Stat-Ease, Inc.}
+  url = {https://github.com/statease/dexpy},
+  urldate = {2025-05-05}
 }
 
 @misc{dirkaudaz_dirkaudaz_2018,
@@ -195,6 +196,7 @@ url={https://openreview.net/forum?id=U1f6wHtG1g}
   author = {{dirkaudaz}},
   year = {2018},
   month = apr,
+  url = {https://github.com/dirkaudaz/DOEPy},
   urldate = {2025-05-05},
   copyright = {BSD-3-Clause}
 }
@@ -203,6 +205,7 @@ url={https://openreview.net/forum?id=U1f6wHtG1g}
   title = {Danieleongari/Definitive\_screening\_design},
   author = {Ongari, Daniele},
   year = {2024},
+  url = {https://github.com/danieleongari/definitive_screening_design},
   month = oct,
   urldate = {2025-05-05},
   copyright = {BSD-3-Clause}
