### 2025-11-21 #8510 @jbytecode
@colinbousige - We have some minor issues:

- The license in Zenodo doesn't match the license in the repository (should change to MIT in Zenodo)
- The metadata including author list, orcids, and title should match those in the paper (The title doesn't match).

Please fix and ping me.
