### 2025-11-21 de37c16
updated to display JOSS paper reference
