### 2025-11-21 #8510 @colinbousige
Gotta say I'm not a big user of social media :-) You see, you actually lead me to make my first LinkedIn post ! Thanks !
